pyproject.toml
[project]
name = "history-issue-features"
version = "0.1.0"
description = "Extract SV features from Verilator/CIRCT historical issues/PRs"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]
